refactor(network): group transport protocol modules
Move the TCP, UDP, and ICMP implementations under protocol-specific module trees and update internal imports to use the new canonical paths. Preserve the existing root module paths as re-exports so downstream users are not broken.
