+++ crates/network/lib/dns/proxies/udp.rs
@@ -20,7 +20,7 @@ use super::super::common::transport::Transport;
 use super::super::forwarder::{DnsForwarder, DnsForwarderHandle};
 use super::super::interceptor::DnsQuery;
 use crate::shared::SharedState;
-use crate::udp_relay::construct_udp_response;
+use crate::udp::relay::construct_udp_response;
 
 //--------------------------------------------------------------------------------------------------
 // Constants
