+++ crates/network/lib/icmp/mod.rs
@@ -1,0 +1,4 @@
+//! ICMP error construction and host-side echo relay.
+
+pub mod error;
+pub mod relay;
