+++ crates/network/lib/lib.rs
@@ -16,24 +16,33 @@ mod addr;
 pub mod backend;
 pub mod builder;
 pub mod config;
-pub mod conn;
 pub mod device;
 pub mod dns;
-pub mod icmp_error;
-pub mod icmp_relay;
+pub mod icmp;
 pub mod network;
 pub mod policy;
-pub mod proxy;
 pub mod publisher;
 pub mod secrets;
 pub mod shared;
 pub mod stack;
+pub mod tcp;
 pub mod tls;
-pub mod udp_fragments;
-pub mod udp_relay;
+pub mod udp;
+
+//--------------------------------------------------------------------------------------------------
+// Constants
+//--------------------------------------------------------------------------------------------------
 
 /// Static hostname the guest uses to reach the sandbox host.
 ///
 /// The host-side DNS interceptor matches guest queries against this
 /// name, and agentd writes the same name into `/etc/hosts`.
 pub(crate) const HOST_ALIAS: &str = "host.microsandbox.internal";
+
+//--------------------------------------------------------------------------------------------------
+// Re-Exports
+//--------------------------------------------------------------------------------------------------
+
+pub use icmp::{error as icmp_error, relay as icmp_relay};
+pub use tcp::{connection as conn, proxy};
+pub use udp::{fragments as udp_fragments, relay as udp_relay};
