+++ crates/network/lib/tcp/mod.rs
@@ -1,0 +1,5 @@
+//! TCP connection tracking, proxying, and host-side upstream dialing.
+
+pub mod connection;
+pub mod proxy;
+pub(crate) mod upstream;
