+++ crates/network/lib/tcp/upstream.rs
@@ -5,7 +5,7 @@ use std::net::SocketAddr;
 
 use tokio::net::TcpStream;
 
-use crate::conn::ProxyConnectState;
+use super::connection::ProxyConnectState;
 use crate::shared::SharedState;
 
 //--------------------------------------------------------------------------------------------------
@@ -118,8 +118,8 @@ fn fallback_eligible(error: &io::Error) -> bool {
 mod tests {
     use tokio::net::TcpListener;
 
+    use super::super::connection::ProxyConnectStatus;
     use super::*;
-    use crate::conn::ProxyConnectStatus;
 
     #[tokio::test]
     async fn connect_falls_back_from_ipv6_to_ipv4_loopback() {
