+++ crates/network/lib/udp/mod.rs
@@ -1,0 +1,4 @@
+//! UDP fragment reassembly and host-side datagram relay.
+
+pub mod fragments;
+pub mod relay;
